% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDriors.R
\name{addDriors}
\alias{addDriors}
\title{Add Driors}
\usage{
addDriors(stocks, priors, same.priors = stocks.combined, stocks.combined,
  shape_prior = 2, b_ref_type = "k", growth_rate_prior = NA,
  growth_rate_prior_cv = 0.2, ...)
}
\arguments{
\item{stocks}{\code{tibble} containing \code{stock}, \code{taxa}, and
\code{data}.}

\item{priors}{data frame containing \code{stock}, \code{initial_state},
\code{initial_state_cv}, \code{terminal_state}, and
\code{terminal_state_cv}.}

\item{same.priors}{whether to use the same priors for all stocks.}

\item{stocks.combined}{old name for \code{same.prior} provided for backward
compatibility. Use \code{same.prior} in new SOFIA scripts.}

\item{shape_prior}{passed to \code{format_driors}.}

\item{b_ref_type}{passed to \code{format_driors}.}

\item{growth_rate_prior}{passed to \code{format_driors}.}

\item{growth_rate_prior_cv}{passed to \code{format_driors}.}

\item{\dots}{additional arguments passed to \code{format_driors}.}
}
\value{
A \code{tibble} like \code{stocks} but with the additional column
\code{driors}.
}
\description{
Add driors (data and priors) column to \sQuote{stocks} object.
}
\details{
If \code{same.priors = TRUE} then the priors where \code{stock = All} are
used for all stocks.

If \code{same.priors = FALSE} then the priors are read from stock-specific
rows in the \code{priors} data frame.
}
\examples{
\dontrun{
addDriors(stocks, priors, same.priors=TRUE)
addDriors(stocks, priors, same.priors=FALSE)
}

}
\seealso{
\code{\link{format_driors}} is the underlying function used to construct the
\code{driors} column.

\code{\link{addEffort}} adds an effort column to catch data.

\code{\link{SOFIA-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
