% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEffort.R
\name{addEffort}
\alias{addEffort}
\title{Add Effort}
\usage{
addEffort(catch, effort, same.effort)
}
\arguments{
\item{catch}{\code{tibble} containing \code{year}, \code{stock}, and
\code{capture}.}

\item{effort}{\code{tibble} containing \code{year}, \code{stock}, and
\code{effort}.}

\item{same.effort}{whether to use the same effort data for all stocks.}
}
\value{
Data frame like \code{catch} but with the additional column \code{effort}.
}
\description{
Add effort column to catch data.
}
\details{
If \code{same.effort = TRUE} then the effort data for \code{stock$All} is
used for all stocks.

If \code{same.effort = FALSE} then the catch and effort data are paired by
year and stock.
}
\examples{
\dontrun{
addEffort(catch, effort, same.effort=TRUE)
addEffort(catch, effort, same.effort=FALSE)
}

}
\seealso{
\code{\link{merge}} is the underlying function used to merge the two data
frames.

\code{\link{addDriors}} adds a driors column to stocks.

\code{\link{SOFIA-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
