% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIndex.R
\name{addIndex}
\alias{addIndex}
\title{Add Index}
\usage{
addIndex(catch, index, same.index)
}
\arguments{
\item{catch}{\code{tibble} containing \code{year}, \code{stock}, and
\code{capture}.}

\item{index}{\code{tibble} containing \code{year}, \code{stock}, and
\code{index}.}

\item{same.index}{whether to use the same index data for all stocks.}
}
\value{
Data frame like \code{catch} but with the additional column \code{index}.
}
\description{
Add index column to catch data.
}
\details{
If \code{same.index = TRUE} then the index data for \code{stock$All} is used
for all stocks.

If \code{same.index = FALSE} then the catch and index data are paired by year
and stock.
}
\examples{
\dontrun{
addIndex(catch, index, same.index=TRUE)
addIndex(catch, index, same.index=FALSE)
}

}
\seealso{
\code{\link{merge}} is the underlying function used to merge the two data
frames.

\code{\link{addEffort}} adds an effort column to catch data.

\code{\link{addDriors}} adds a driors column to stocks.

\code{\link{SOFIA-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
