% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compCat.R
\name{compCat}
\alias{compCat}
\title{Compute Categories}
\usage{
compCat(dat, method = "cmsy.naive")
}
\arguments{
\item{dat}{a data frame containing columns named \code{bbmsy.*} and
\code{ffmsy.*}.}

\item{method}{a string indicating which method was used to estimate B/Bmsy
and F/Fmsy.}
}
\value{
Data frame like \code{dat} but with additional columns containing stock
status category information.
}
\description{
Calculate stock status categories from B/Bmsy and F/Fmsy, and add as columns
to an existing data frame.
}
\details{
The column names in \code{dat} should contain the \code{method} name as a
suffix. For example, if \code{method = "effEdepP"}, then this function will
look for columns called \code{bbmsy.effEdepP} and \code{ffmsy.effEdepP}.
}
\examples{
\dontrun{
compCat(newResTab, method="effDepP")
}

}
\seealso{
\code{\link{plotProp}} plots a summary of stock status categories.

\code{\link{TSAF-package}} gives an overview of the package.
}
\author{
Dan Ovando, with a contribution by Arni Magnusson.
}
