% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertData.R
\name{convertData}
\alias{convertData}
\title{Convert Data}
\usage{
convertData(subdir, min.catch = 0.001, quiet = FALSE)
}
\arguments{
\item{subdir}{directory containing the primary data files.}

\item{min.catch}{small numeric constant to use instead of \code{NA} or zero
catch.}

\item{quiet}{whether to suppress screen output.}
}
\value{
Files are converted inside \code{subdir}. As a byproduct, a list is returned,
describing the files after conversion.
}
\description{
Convert primary single-stock data files into combined data files ready for
analysis, such \code{catch.csv}, \code{effort.csv}, and/or \code{index.csv}.
}
\details{
\code{subdir} is usually one of \code{"both"}, \code{"effort"},
\code{"index"}, or \code{"neither"}, generated by the \code{groupData}
function.

Pass any non-numeric value for \code{min.catch}, for example \code{FALSE}, to
retain \code{NA} and zero catch values.
}
\note{
A primary data file can have a filename such as
\file{Yellowtail_snapper_Mexico.csv} and columns such as
\code{stockid|scientificname|commonname|year|catch|stocklong}.

The combined data files, such as \code{catch.csv}, have the format
\code{Year|Unique_Stock_Name_1|Unique_Stock_Name_2|...} where, for example,
\code{Yellowtail_snapper_Mexico} is one column name.

The output files created will depend on whether the data include effort
and/or index data. If \code{subdir = "both"}, a full set of output files will
be created: \code{catch.csv}, \code{effort.csv}, and \code{index.csv}.

The functions \code{groupData} and \code{convertData} are used together:
first group, then convert.
}
\examples{
\dontrun{
convertData("Data_files_Area_31_3/both")
convertData("Data_files_Area_31_3/effort, quiet=TRUE")
}

}
\seealso{
\code{\link{groupData}} groups primary data in subdirectories.

\code{\link{SOFIA-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
