% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitCloneAll.R
\name{gitCloneAll}
\alias{gitCloneAll}
\title{gitCloneAll}
\usage{
gitCloneAll(topdir = "c:/git/sofia-taf", method = "https", tree = TRUE)
}
\arguments{
\item{topdir}{local empty directory to clone to.}

\item{method}{cloning method, either \code{"https"} or \code{"ssh"}.}

\item{tree}{whether to organize repositories in a nested SOFIA tree, see
details below.}
}
\value{
String vector of cloned repositories.
}
\description{
Clone all SOFIA-TAF repositories from GitHub.
}
\details{
Generally, the default \code{method = "https"} works best in Windows, while
\code{method = "ssh"} works best in Linux.

When \code{tree = TRUE}, this function will recognize standard SOFIA
repository names such as \code{2022Area37Demo} and clone to a nested
subdirectory \code{2022/Area37/Demo} instead of \code{2022Area37Demo}.
}
\examples{
\dontrun{
gitCloneAll()
gitCloneAll("~/git/sofia-taf", "ssh")
}

}
\seealso{
\code{\link{gitRepos}} fetches a complete list of GitHub repositories.

\code{\link{gitClone}} clones a repository from GitHub.

\code{\link{SOFIA-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
