% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCatchEffort.R
\name{mergeCatchEffort}
\alias{mergeCatchEffort}
\title{Merge Catch and Effort}
\usage{
mergeCatchEffort(catch, effort, stocks.combined)
}
\arguments{
\item{catch}{data frame (or \code{tibble}) containing \code{year},
\code{stock}, and \code{capture}.}

\item{effort}{data frame (\code{tibble}) containing \code{year},
\code{stock}, and \code{effort}.}

\item{stocks.combined}{whether to use the same effort data for all stocks.}
}
\value{
Data frame like \code{catch} but with the additional column \code{effort}.
}
\description{
Merge catch and effort data into a common data frame.
}
\details{
If \code{stocks.combined = TRUE} then the effort data for \code{stock$All} is
used for all stocks.

If \code{stocks.combined = FALSE} then the catch and effort data are paired
by year and stock.
}
\examples{
\dontrun{
mergeCatchEffort(catch, effort, stocks.combined=TRUE)
mergeCatchEffort(catch, effort, stocks.combined=FALSE)
}

}
\seealso{
\code{\link{merge}} is the underlying function used to merge the two data
frames.

\code{\link{TSAF-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
