% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCat.R
\name{plotCat}
\alias{plotCat}
\alias{plotProp}
\title{Plot Categories}
\usage{
plotCat(dat, method = "cmsy.naive", cats = 4, type = "count")
}
\arguments{
\item{dat}{data frame of stock time series, containing columns named
\code{bbmsy} and \code{ffmsy}, as well as method-specific
\code{bbmsy.*} and \code{ffmsy.*}.}

\item{method}{string indicating which method was used to estimate B/Bmsy and
F/Fmsy.}

\item{cats}{either \code{3} or \code{4}, indicating whether to plot the stock
status based on biomass only (3 categories) or based on biomass and
fishing mortality (4 categories).}

\item{type}{string indicating the type of plot, either \code{"count"} or
\code{"stock"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a summary of stock status categories by year: underfished (green), fully
fished (yellow), or overfished (red).
}
\details{
The first two columns in \code{dat} are treated as \code{'stock'} and
\code{'year'}, regardless of the actual column names.

The last column names in \code{dat} should contain the \code{method} name as
a suffix. For example, if \code{method = "effEdepP"}, then this function will
look for columns called \code{bbmsy.effEdepP} and \code{ffmsy.effEdepP}.

The \code{type} argument supports legacy synonyms, where \code{"prop"} is the
same as \code{"count"}, and \code{"all"} is the same as \code{"stock"}.
}
\examples{
\dontrun{
plotCat(newResTab, method="effEdepP", cats=3, type="count")
plotCat(newResTab, method="effEdepP", cats=3, type="stock")
}

}
\seealso{
\code{\link{ggplot}} is the underlying function used to produce the plot.

\code{\link{calcCat}} calculates stock status categories.

\code{\link{SOFIA-package}} gives an overview of the package.
}
\author{
Rishi Sharma and Arni Magnusson.
}
