% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProp.R
\name{plotProp}
\alias{plotProp}
\title{Plot Proportions}
\usage{
plotProp(dat, method = "cmsy.naive", cats = 4, type = "prop")
}
\arguments{
\item{dat}{a data frame of stock time series, containing B/Bmsy and F/Fmsy.}

\item{method}{a string indicating which method was used to estimate B/Bmsy
and F/Fmsy.}

\item{cats}{either \code{3} or \code{4}, indicating whether to plot the stock
status based on biomass only (3 categories) or based on biomass and
fishing mortality (4 categories).}

\item{type}{a string indicating the type of plot, either \code{"prop"} or
\code{"all"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a summary of stock status categories by year: underfished (green), fully
fished (yellow), or overfished (red).
}
\examples{
\dontrun{
plotProp(newResTab, method="effEdepP", cats=3, type="prop")
plotProp(newResTab, method="effEdepP", cats=3, type="all")
}

}
\seealso{
\code{\link{ggplot}} is the underlying function used to produce the plot.

\code{\link{compCat}} calculates stock status categories.

\code{\link{TSAF-package}} gives an overview of the package.
}
\author{
Dan Ovando, with a contribution by Arni Magnusson.
}
