% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCatch.R
\name{readCatch}
\alias{readCatch}
\title{Read Catch}
\usage{
readCatch(file = "bootstrap/data/catch.csv", stocks.combined)
}
\arguments{
\item{file}{filename, \code{"bootstrap/data/catch.csv"} by default.}

\item{stocks.combined}{whether stocks are combined in the subsequent
analysis.}
}
\value{
Data frame containing year in the first column and catch in other columns.
}
\description{
Read catch data and handle the \code{Total} column appropriately.
}
\details{
If \code{stocks.combined = TRUE} then only the \code{Total} column is used
and renamed to \code{All}. This way, it will match the \code{All} column
found in the effort data and priors.

If \code{stocks.combined = FALSE} then the \code{Total} column is discarded.
}
\examples{
\dontrun{
readCatch("bootstrap/data/catch.csv", stocks.combined=TRUE)
readCatch(stocks.combined=FALSE)
}

}
\seealso{
\code{\link{read.csv}} is the underlying function used to read the catch
data.

\code{\link{TSAF-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
