% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofiaColors.R
\docType{data}
\name{sofiaColors}
\alias{sofiaColors}
\alias{sofiaDarkblue}
\alias{sofiaBlue}
\alias{sofiaOrange}
\title{SOFIA Colors}
\usage{
sofiaDarkblue
sofiaBlue
sofiaOrange
sofiaColors
}
\description{
Predefined colors that can be useful in SOFIA plots.
}
\note{
The SOFIA 2024 report uses #1179be (blue) and #f07e29 (orange).

The \code{sofiaDarkblue} and \code{sofiaBlue} colors offer two hues of blue
that can be used to distinguish between fully and underfished.
}
\examples{
sofiaBlue

barplot(rep(1,3), space=0, horiz=TRUE, col=sofiaColors, border=NA)
}
\seealso{
\code{\link{SOFIA-package}} gives an overview of the package.
}
