% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransparentSOFIA-package.R
\docType{package}
\name{TransparentSOFIA-package}
\alias{TransparentSOFIA-package}
\alias{TransparentSOFIA}
\title{Tools to Work with SOFIA Analyses}
\description{
Tools that support the Transparent SOFIA framework.
}
\details{
\emph{Prepare data:}
\tabular{ll}{
  \code{\link{addDriors}} \tab add driors column to stocks object\cr
  \code{\link{addEffort}} \tab add effort column to catch data
}
\emph{Calculate:}
\tabular{ll}{
  \code{\link{calcCat}} \tab stock status categories
}
\emph{Plot:}
\tabular{ll}{
  \code{\link{plotCat}} \tab summary of stock status categories
}
}
\references{
\url{https://github.com/sofia-tsaf/TransparentSOFIA}
}
\seealso{
\pkg{TransparentSOFIA} is used in combination with the packages \pkg{sraplus}
(analytical model) and \pkg{TAF} (workflow).
}
\author{
Rishi Sharma and Arni Magnusson.
}
