% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSAF-package.R
\docType{package}
\name{TSAF-package}
\alias{TSAF-package}
\alias{TSAF}
\title{Tools to Work with SOFIA-TSAF Analyses}
\description{
Tools that support the Transparent SOFIA Assessment Framework (TSAF)
analyses.
}
\details{
\emph{Prepare data:}
\tabular{ll}{
  \code{\link{addDriors}} \tab add driors column to stocks object\cr
  \code{\link{addEffort}} \tab add effort column to catch data
}
\emph{Calculate:}
\tabular{ll}{
  \code{\link{compCat}} \tab stock status categories
}
\emph{Plot:}
\tabular{ll}{
  \code{\link{plotProp}} \tab summary of stock status categories
}
}
\references{
\url{https://github.com/sofia-tsaf/TSAF}
}
\seealso{
\pkg{TSAF} is used in combination with the packages \pkg{sraplus} (analytical
model) and \pkg{TAF} (workflow).
}
\author{
Rishi Sharma and Arni Magnusson.
}
