% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCat.R
\name{calcCat}
\alias{calcCat}
\alias{compCat}
\title{Calculate Categories}
\usage{
calcCat(dat, method = "cmsy.naive")
}
\arguments{
\item{dat}{data frame of stock time series, containing columns named
\code{bbmsy} and \code{ffmsy}, as well as method-specific
\code{bbmsy.*} and \code{ffmsy.*}.}

\item{method}{string indicating which method was used to estimate B/Bmsy and
F/Fmsy.}
}
\value{
Data frame like \code{dat} but with additional columns containing stock
status category information.
}
\description{
Calculate stock status categories from B/Bmsy and F/Fmsy, and add as columns
to an existing data frame.
}
\details{
The column names in \code{dat} should contain the \code{method} name as a
suffix. For example, if \code{method = "effEdepP"}, then this function will
look for columns called \code{bbmsy.effEdepP} and \code{ffmsy.effEdepP}.
}
\examples{
\dontrun{
calcCat(newResTab, method="effDepP")
}

}
\seealso{
\code{\link{plotCat}} plots a summary of stock status categories.

\code{\link{TSAF-package}} gives an overview of the package.
}
\author{
Rishi Sharma, with a contribution by Arni Magnusson.
}
