% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitClone.R
\name{gitClone}
\alias{gitClone}
\title{gitClone}
\usage{
gitClone(repo, topdir = "c:/git/sofia-taf", tree = TRUE,
  method = "https")
}
\arguments{
\item{repo}{GitHub repository of the form \code{"repo"} or
\code{"owner/repo"}.}

\item{topdir}{local directory above the repository.}

\item{tree}{whether to organize repositories in a nested SOFIA tree, see
details below.}

\item{method}{cloning method, either \code{"https"} or \code{"ssh"}.}
}
\value{
String containing the cloning command.
}
\description{
Clone a repository from GitHub.
}
\details{
The implicit default \code{repo} owner is \code{sofia-taf}. In other words,
if the user specifies a \code{"repo"} instead of \code{"owner/repo"}, this
function will automatically convert it to \code{"sofia-taf/repo"}.

Generally, the default \code{method = "https"} works best in Windows, while
\code{method = "ssh"} works best in Linux.

When \code{tree = TRUE}, this function will recognize standard SOFIA
repository names such as \code{2022Area37Demo} and clone to a nested
subdirectory \code{2022/Area37/Demo} instead of \code{2022Area37Demo}. For
repository names that do not follow this YearAreaAnalysis pattern, the value
of \code{tree} has no effect.
}
\examples{
\dontrun{
gitClone("2022Area37Demo")
gitClone("arni-magnusson/areaplot", "~/git/arni-magnusson", method="ssh")
}

}
\seealso{
\code{\link{gitRepos}} fetches a complete list of GitHub repositories.

\code{\link{SOFIA-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
