% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSAF-package.R
\docType{package}
\name{TSAF-package}
\alias{TSAF-package}
\alias{TSAF}
\title{Tools to Work with SOFIA-TSAF Analyses}
\description{
Tools that support the Transparent SOFIA Assessment Framework (TSAF)
analyses.
}
\details{
\emph{Calculate:}
\tabular{ll}{
  \code{\link{compCat}} \tab stock status categories
}
}
\references{
\url{https://github.com/sofia-tsaf/TSAF}
}
\seealso{
\pkg{TSAF} is used in combination with the packages \pkg{sraplus} (analytical
model) and \pkg{TAF} (workflow).
}
\author{
Dan Ovando, Arni Magnusson, and Rishi Sharma.
}
